Updates calon murid list and create user functionalities
Updates the calon murid list query to use the correct column names and table.

Also, it copies the registration data to the `formulir_pendaftaran` table when creating a user,
and disables actions on the calon murid list if the student's status is "approved" and a user has been created.
This prevents accidental modifications after user creation.