Connects registrations to formulir pendaftaran
Adds a foreign key to the registrations table referencing the formulir_pendaftaran table.

This change allows linking registration records to their corresponding formulir_pendaftaran entries. It also updates the create-user API to populate the formulir_pendaftaran_id in the registrations table and modifies the student-portfolio API to query the registrations table instead of calon_murid.